+++ CHANGES.md
@@ -1,3 +1,9 @@
+# Change Log
+
+## 0.1.3
+* added success message type
+* fix editing-to-add-DONE bug
+* fixed backup deletion bug
 ## 0.1.2 *2020-1-23*
 * adding more fractional precision options: half hour, third, etc.
 ## 0.1.1 *2020-1-20*